fix(viewer): stop file:// history replay from self-triggering an init loop
Reopening a recording via remembered File System Access permission
(?url=file://...) could self-trigger forever: replaying history set
activeId, which changed appendFilesAsDatasets's identity, which re-ran
the SPA bootstrap effect that depends on it, which replayed history
again. Symptoms were a flickering sidebar/player and hundreds of
WorkerSourceCancelledError as the player was torn down and rebuilt in
a tight loop, occasionally crashing the tab. Direct drag-and-drop was
unaffected since it's a one-shot event handler, not an effect.

Root-caused and fixed at the design-pattern level rather than patched:
- useSpaUrlBootstrap wraps its handlers in React's useEffectEvent, so
  its effect only ever depends on (active, url) - no callback identity
  churn can retrigger it. react-hooks/exhaustive-deps now errors if
  this guarantee is ever violated, turning the bug class into a lint
  error instead of a silent regression.
- resolveAppendGroupId/resolveActiveId (datasetSources.ts) make the
  "which group is really active" logic pure and directly testable,
  fixing the underlying activeId "orphan id" bounce.
- Regression tests pin both the pure logic and the hook's Effect
  timing (useSpaUrlBootstrap.test.tsx feeds in a worse-than-production
  case: a fresh handler identity on every render).

Also splits the 1400+ line RosViewerImpl.tsx into focused, independently
testable hooks/components (useDatasetSession, usePlayerLifecycle,
useRecordingSourceActions, useDatasetHistory, useThemeLanguagePreferences,
useOpenFeedback, RosViewerLandingScreen, sourceLoading, rosViewerUtils),
leaving RosViewerImpl.tsx as a thin composition layer. Fixed a latent bug
found during the split: the hidden #rosview-inline-file/-tar inputs must
render unconditionally even once a player exists, or history replay can't
click them.
